+++ R/slicejam-piecharts.R
@@ -25,7 +25,7 @@ feature_to_pie <- function
  colorSub=NULL,
  ...)
 {
-   j <- table(rmNA(x));
+   j <- table(jamba::rmNA(x));
    if (length(j) == 0) {
       return(NULL);
    }
@@ -50,9 +50,9 @@ feature_to_pie <- function
       main <- NULL
    }
    j_names <- names(j);
-   colorSub <- rmNA(colorSub[j_names]);
+   colorSub <- jamba::rmNA(colorSub[j_names]);
    if (length(colorSub) == length(j_names)) {
-      colorSub <- nameVector(rainbowJam(length(j_names)),
+      colorSub <- jamba::nameVector(rainbowJam(length(j_names)),
          j_names);
    }
    names(j) <- paste0(j_names,
@@ -182,16 +182,17 @@ peak_pie_by_region <- function
 #' @family slicejam plot functions
 #' 
 #' @param hit_array `array` output from `se_contrast_stats()` with
-#'    three dimensions: hit, contrasts, signal. The arguments
-#'    `iHit`, `iCon`, and `iSig` are used to subset the array for
-#'    a specific result to use in each pie chart. Each cell of the
-#'    array is expected to contain a numeric list with values `c(-1, 1)`
-#'    with names that are peak names.
+#'    three dimensions: Cutoffs, Contrasts, Signal.
+#'    The arguments `iHit`, `iCon`, and `iSig` are used to subset
+#'    the array to define the specific cell corresponding to a Contrast,
+#'    Hit criteria, and Signal.
+#'    Each cell of the array is expected to contain a numeric list
+#'    with values `c(-1, 1)` using names that are peak names.
 #' @param feature_type_winner `character` or `factor` vector whose names
 #'    are expected to be peak names, which are also names of entries
-#'    in `hit_array`.
-#' @param iHit,iCon,iSig `character` value indicating a specific hit,
-#'    contrast, signal, respectively.
+#'    in `hit_array`. The values will become labels in the pie chart.
+#' @param iHit,iCon,iSig one `character` string for each argument,
+#'    indicating a specific Hit, Contrast, and Signal, respectively.
 #' @param iPeaks `character` not used in this implementation.
 #' @param do_pie `logical` indicating whether to return the pie chart,
 #'    or when `do_pie=FALSE` the `data.frame` is returned.
@@ -199,10 +200,32 @@ peak_pie_by_region <- function
 #' @param label_type `character` indicating the type of pie chart label
 #'    to place around the outside: `"text"` uses `ggrepel::geom_text_repel()`,
 #'    and `"label"` uses `ggrepel::geom_label_repel()` with colored background.
+#' @param label_size `numeric` relative label size.
+#' @param base_size `numeric` font size passed to `colorjam::theme_jam()`,
+#'    in terms of base fontsize in the `ggplot2` theme.
+#' @param strip_size `numeric` adjustment for the `ggplot2` facet strip
+#'    size above each plot panel.
+#' @param strip_sub_size `numeric` font size for the facet strip text.
+#' @param summary_type `character` string indicating the summary measurement:
+#'    * `"peak"` (default) tabulates the number of peaks per feature type.
+#'    * `"width"` is available when `feature_type` is provided as `GRanges`,
+#'    and tabulates the number of bases covered, total width, per
+#'    feature type.
+#' @param verbose `logical` indicating whether to print verbose output.
+#' @param ... additional arguments are ignored.
+#' 
+#' @examples
+#' countset <- jamba::nameVector(c(978576, 327871, 127049, 188781, 29365),
+#'    c("Promoters (-2000,+200)", "TTS (-1000,+1000)",
+#'    "exons", "introns", "extragenic"))
+#' ft_data <- rep(names(countset), countset);
+#' names(ft_data) <- paste0("peak", seq_along(ft_data));
+#' gg_pie_by_feature_type(feature_type_winner=ft_data,
+#'    add_peak_list=list(`Summary`=names(ft_data)))
 #' 
 #' @export
 gg_pie_by_feature_type <- function
-(hit_array,
+(hit_array=NULL,
  feature_type_winner,
  PeakSet="",
  iHit=NULL,
@@ -213,17 +236,64 @@ gg_pie_by_feature_type <- function
  do_pie=TRUE,
  main=NULL,
  colorSub=NULL,
- label_type=c("text", "label"),
- label_size=5,
- base_size=16,
+ label_type=c("text",
+    "label"),
+ label_size=4,
+ base_size=14,
  strip_size=24,
  strip_sub_size=12,
+ summary_type=c("peak",
+    "width"),
+ verbose=FALSE,
  ...)
 {
    #
    label_type <- match.arg(label_type);
+   summary_type <- match.arg(summary_type);
+   if ("width" %in% summary_type) {
+      if (!inherits(feature_type_winner, "GRanges")) {
+         stop("when summary_type='width', feature_type_winner must be GRanges")
+      }
+   }
+   use_gr <- NULL;
+   feature_widths <- NULL;
+   if (inherits(feature_type_winner, "GRanges")) {
+      if (verbose) {
+         jamba::printDebug("gg_pie_by_feature_type(): ",
+            "Recognized feature_type_winner as GRanges.");
+      }
+      use_gr <- feature_type_winner;
+      feature_type_winner <- jamba::nameVector(
+         GenomicRanges::values(use_gr)$feature_type_winner,
+         names(use_gr));
+      # feature_widths
+      if ("width" %in% summary_type) {
+         if (verbose) {
+            jamba::printDebug("gg_pie_by_feature_type(): ",
+               "Defined feature_widths.");
+         }
+         feature_widths <- jamba::nameVector(
+            GenomicRanges::width(use_gr),
+            names(use_gr));
+      }
+   }
+   
    if (length(feature_type_winner) == 0) {
       stop("feature_type_winner must be supplied and non-empty.")
+   }
+   
+   # enforce factor order for consistency later
+   if (!is.factor(feature_type_winner)) {
+      if (verbose) {
+         jamba::printDebug("gg_pie_by_feature_type(): ",
+            "Converted feature_type_winner to factor.");
+      }
+      feature_type_winner_levels <- jamba::provigrep(
+         c("promoter", "tts", "exon", "intron",
+            "extragenic", "intragenic", ".*"),
+         unique(feature_type_winner));
+      feature_type_winner <- factor(feature_type_winner,
+         levels=feature_type_winner_levels);
    }
    
    if (length(iHit) == 0) {
@@ -242,14 +312,31 @@ gg_pie_by_feature_type <- function
       iCon <- dimnames(hit_array)[[2]][iSig];
    }
    
-   peak_to_tabledf <- function(k, feature_type_winner, iCon, iHit, iSig)
+   peak_to_tabledf <- function
+   (k,
+    feature_type_winner,
+    iCon,
+    iHit,
+    iSig,
+    feature_widths=NULL)
    {
-      j <- table(rmNA(feature_type_winner[k]));
+      # basic peak count
+      use_label <- "peaks";
+      if (length(feature_widths) == 0) {
+         j <- table(jamba::rmNA(feature_type_winner[k]));
+      } else {
+         k_list <- split(k, feature_type_winner[k])
+         j <- sapply(k_list, function(k1){
+            sum(feature_widths[k1])
+         })
+         names(j) <- names(k_list);
+         use_label <- "bases";
+      }
       j_pct <- as.integer(j) / sum(j) * 100;
       j_df <- data.frame(check.names=FALSE,
          stringsAsFactors=FALSE,
-         feature_type=names(j),
-         Count=as.integer(j),
+         feature_type=factor(names(j), levels=unique(names(j))),
+         Count=as.numeric(as.integer(j)),
          Scaled=j_pct,
          Percent=paste0(
             format(j_pct,
@@ -257,20 +344,24 @@ gg_pie_by_feature_type <- function
                trim=TRUE),
             "%")
       );
-      j_df$Offset <- sum(j_df$Count) -(cumsum(j_df$Count) - j_df$Count / 2);
+      j_df$Offset <- sum(j_df$Count) -
+         (cumsum(as.numeric(j_df$Count)) - j_df$Count / 2);
       j_df$Scaled_Offset <- 100 * j_df$Offset / sum(j_df$Count, na.rm=TRUE);
       
       # change to use comp, an abbreviated format
       # j_df$Set <- iCon;
       j_df$Set <- jamses::contrast2comp(iCon);
       
+      use_labels <- ifelse(j_df$Count == 1,
+         gsub("s^", "", use_label),
+         use_label)
       j_df$Hit <- iHit;
       j_df$Signal <- iSig;
       j_df$Label <- paste0(
          j_df$feature_type,
          "\n",
          jamba::formatInt(j_df$Count),
-         " peaks\n",
+         " ", use_labels, "\n",
          j_df$Percent);
       j_df;
    }
@@ -282,17 +373,16 @@ gg_pie_by_feature_type <- function
             j_df <- NULL;
             if (length(k) > 0) {
                j_df <- peak_to_tabledf(k,
-                  feature_type_winner,
+                  feature_type_winner=feature_type_winner,
                   iCon1,
                   iHit1,
-                  iSig1);
+                  iSig1,
+                  feature_lengths=feature_lengths);
             }
             j_df;
          }))
       }))
    }));
-   #jamba::printDebug("head(jdf, 10):");
-   #print(head(jdf, 10));
    
    # optional extra peaks
    if (length(add_peak_list) > 0) {
@@ -302,10 +392,11 @@ gg_pie_by_feature_type <- function
             k <- names(k)
          }
          j_df <- peak_to_tabledf(k,
-            feature_type_winner,
+            feature_type_winner=feature_type_winner,
             iCon=iname,
             iHit="",
-            iSig="");
+            iSig="",
+            feature_widths=feature_widths);
       });
       jdf_add <- jamba::rbindList(jdf_add_list);
       #jamba::printDebug("head(jdf_add, 10):");
@@ -314,6 +405,8 @@ gg_pie_by_feature_type <- function
    }
    jdf$Set <- factor(jdf$Set,
       levels=unique(jdf$Set));
+   jdf$Label <- factor(jdf$Label,
+      levels=unique(jdf$Label));
    jdf$Hit <- factor(jdf$Hit,
       levels=unique(jdf$Hit));
    jdf$Signal <- factor(jdf$Signal,
@@ -333,7 +426,7 @@ gg_pie_by_feature_type <- function
    jdf$Panel <- jamba::pasteByRowOrdered(
       jdf[,c("Set", "Hit", "Signal")],
       sep="|");
-   
+
    if (!do_pie) {
       attr(jdf, "panels") <- levels(jdf$Panel);
       attr(jdf, "panel_count") <- length(levels(jdf$Panel));
@@ -376,6 +469,10 @@ gg_pie_by_feature_type <- function
                useGrey=7))
    }
    
+   # custom legend title
+   gg2 <- gg2 +
+      ggplot2::guides(fill=ggplot2::guide_legend(title="Feature Type"))
+   
    # add text information around the pie chart
    if ("text" %in% label_type) {
       gg2 <- gg2 +
@@ -389,13 +486,15 @@ gg_pie_by_feature_type <- function
             size=label_size,
             show.legend=FALSE)
    } else if ("label" %in% label_type) {
-      gg2 +
+      gg2 <- gg2 +
          ggrepel::geom_label_repel(
             segment.color="black",
-            aes(x=1.525,
-               color=feature_type,
+            color="black",
+            fill="white",
+            ggplot2::aes(x=1.525,
+               # color=feature_type,
+               # fill=feature_type,
                y=Scaled_Offset,
-               fill=feature_type,
                label=Label),
             nudge_x=0.9,
             size=label_size,